+++ agent/api/task/task.go
@@ -309,6 +309,15 @@ func (task *Task) addGPUResource() error {
 	return nil
 }
 
+func (task *Task) isGPUEnabled() bool {
+	for _, association := range task.Associations {
+		if association.Type == GPUAssociationType {
+			return true
+		}
+	}
+	return false
+}
+
 func (task *Task) populateGPUEnvironmentVariables() {
 	for _, container := range task.Containers {
 		if len(container.GPUIDs) > 0 {
@@ -982,7 +991,7 @@ func (task *Task) dockerHostConfig(container *apicontainer.Container, dockerCont
 		Resources:    resources,
 	}
 
-	if task.shouldRequireNvidiaRuntime(container) {
+	if task.isGPUEnabled() && task.shouldRequireNvidiaRuntime(container) {
 		if task.NvidiaRuntime == "" {
 			return nil, &apierrors.HostConfigError{"Runtime is not set for GPU containers"}
 		}